# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'QASA API Documentation'
  description: 'API documentation for QASA Laravel application - invoicing and client management system'
  contact:
    name: 'QASA Support'
    email: support@qasa.sk
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
paths:
...
  /api/v1/reports/eu-sales-list:
    get:
      tags:
        - VatReports
      summary: 'EU sales list (súhrnný výkaz) basis — intra-EU reverse-charged invoices grouped by month and client VAT ID'
      operationId: 4aa595a3a66c1b3e8d787d9cec3bcd34
      parameters:
        -
          name: year
          in: query
          required: true
          schema:
            type: integer
            example: 2026
        -
          name: quarter
          in: query
          required: false
          schema:
            type: integer
            maximum: 4
            minimum: 1
        -
          name: month
          in: query
          required: false
          schema:
            type: integer
            maximum: 12
            minimum: 1
      responses:
        200:
          description: 'EU sales list rows'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      properties:
                        period:
                          type: string
                          example: 2026-01
                        vat_id:
                          type: string
                          example: DE123456789
                        client_name:
                          type: string
                        amount:
                          type: number
                          format: float
                        code:
                          description: '3 = services'
                          type: integer
                          example: 3
                      type: object
                type: object
        401:
          description: Unauthenticated
        422:
          description: 'Validation error'
      security:
        -
          sanctum: []
  /api/v1/reports/vat-control-statement:
    get:
      tags:
        - VatReports
      summary: 'VAT control statement basis (SK: kontrolný výkaz DPH / CZ: kontrolní hlášení) — data preparation only, not a filing'
      operationId: e41295ca52bdf24c617c0f6766b86667
      parameters:
        -
          name: country
          in: query
          required: true
          schema:
            type: string
            enum:
              - SK
              - CZ
        -
          name: year
          in: query
          required: true
          schema:
            type: integer
            example: 2026
        -
          name: quarter
          in: query
          required: false
          schema:
            type: integer
            maximum: 4
            minimum: 1
        -
          name: month
          in: query
          required: false
          schema:
            type: integer
            maximum: 12
            minimum: 1
      responses:
        200:
          description: 'VAT control statement sections'
        401:
          description: Unauthenticated
        422:
          description: 'Validation error or the account is not a VAT payer'
      security:
        -
          sanctum: []
  /api/v1/reports/vat-control-statement/xml:
    get:
      tags:
        - VatReports
      summary: 'VAT control statement XML draft (CZ: DPHKH1 / SK: KVDPH_2025) — draft for review before filing, not a submission-ready document'
      operationId: d6a42cff2c8635bd2f4d50c420ebfb70
      parameters:
        -
          name: country
          in: query
          required: true
          schema:
            type: string
            enum:
              - SK
              - CZ
        -
          name: year
          in: query
          required: true
          schema:
            type: integer
            example: 2026
        -
          name: quarter
          in: query
          required: false
          schema:
            type: integer
            maximum: 4
            minimum: 1
        -
          name: month
          in: query
          required: false
          schema:
            type: integer
            maximum: 12
            minimum: 1
      responses:
        200:
          description: 'XML document'
          content:
            application/xml: []
        401:
          description: Unauthenticated
        422:
          description: 'Validation error or the account is not a VAT payer'
      security:
        -
          sanctum: []
...
components:
...
  securitySchemes:
    sanctum:
      type: apiKey
      description: 'Enter token in format (Bearer <token>)'
      name: Authorization
      in: header
...
security:
  -
    sanctum: []